requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Database
pymongo[srv]==4.6.0

# Environment
python-dotenv==1.0.0

# AI & LangChain
langchain==0.1.0
langchain-core==0.1.0
langchain-google-genai==0.0.6
langgraph==0.0.20
google-generativeai==0.3.2

# HTTP & Utilities
requests==2.31.0
typing-extensions==4.9.0

# Development
ipython==8.18.1

# Testing (optional for now)
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality (optional for now)
black==23.12.1
isort==5.13.2
pylint==3.0.3
mypy==1.7.1
